WIP: rework of responses handling for the commands
The busClient must not be used directly in order to handle properly the
responses accordingly to the orc that send the commands. It opens the
possibilities to handle concurrent commands with several or only one
orc.

This commit is just the first step but most of features should continue
to work as expected. Note that many functionalities are still not tested.

The activities can be run concurrently but it's still experimental (but
necessary).

A lot of module must be reworked in order to improve the readability now
that more arguments are passed to the command handlers.